+++ calliope/preprocess/model_run.py
@@ -41,7 +41,9 @@ _DEFAULT_PALETTE = [
 ]
 
 
-def model_run_from_yaml(model_file, scenario=None, override_dict=None):
+def model_run_from_yaml(
+    model_file, timeseries_dataframes=None, scenario=None, override_dict=None
+):
     """
     Generate processed ModelRun configuration from a
     YAML model configuration file.
@@ -50,9 +52,13 @@ def model_run_from_yaml(model_file, scenario=None, override_dict=None):
     ----------
     model_file : str
         Path to YAML file with model configuration.
+    timeseries_dataframes : dict, optional
+        Dictionary of timeseries dataframes. The keys are strings
+        corresponding to the dataframe names given in the yaml files and
+        the values are dataframes with time series data.
     scenario : str, optional
         Name of scenario to apply. Can either be a named scenario, or a
-        comman-separated list of individual overrides to be combined
+        comma-separated list of individual overrides to be combined
         ad-hoc, e.g. 'my_scenario_name' or 'override1,override2'.
     override_dict : dict or AttrDict, optional
 
@@ -65,11 +71,17 @@ def model_run_from_yaml(model_file, scenario=None, override_dict=None):
     )
 
     return generate_model_run(
-        config_with_overrides, debug_comments, overrides, scenario
+        config_with_overrides,
+        timeseries_dataframes,
+        debug_comments,
+        overrides,
+        scenario,
     )
 
 
-def model_run_from_dict(config_dict, scenario=None, override_dict=None):
+def model_run_from_dict(
+    config_dict, timeseries_dataframes=None, scenario=None, override_dict=None
+):
     """
     Generate processed ModelRun configuration from a
     model configuration dictionary.
@@ -77,9 +89,13 @@ def model_run_from_dict(config_dict, scenario=None, override_dict=None):
     Parameters
     ----------
     config_dict : dict or AttrDict
+    timeseries_dataframes : dict, optional
+        Dictionary of timeseries dataframes. The keys are strings
+        corresponding to the dataframe names given in the yaml files and
+        the values are dataframes with time series data.
     scenario : str, optional
         Name of scenario to apply. Can either be a named scenario, or a
-        comman-separated list of individual overrides to be combined
+        comma-separated list of individual overrides to be combined
         ad-hoc, e.g. 'my_scenario_name' or 'override1,override2'.
     override_dict : dict or AttrDict, optional
 
@@ -95,7 +111,11 @@ def model_run_from_dict(config_dict, scenario=None, override_dict=None):
     )
 
     return generate_model_run(
-        config_with_overrides, debug_comments, overrides, scenario
+        config_with_overrides,
+        timeseries_dataframes,
+        debug_comments,
+        overrides,
+        scenario,
     )
 
 
@@ -140,9 +160,10 @@ def apply_overrides(config, scenario=None, override_dict=None):
     )
 
     # Interpret timeseries_data_path as relative
-    config.model.timeseries_data_path = relative_path(
-        config.config_path, config.model.timeseries_data_path
-    )
+    if "timeseries_data_path" in config.model:
+        config.model.timeseries_data_path = relative_path(
+            config.config_path, config.model.timeseries_data_path
+        )
 
     # FutureWarning: check if config includes an explicit objective cost class.
     # Added in 0.6.4-dev, to be removed in v0.7.0-dev.
@@ -433,7 +454,59 @@ def process_tech_groups(config_model, techs):
     return tech_groups
 
 
-def process_timeseries_data(config_model, model_run):
+def load_timeseries_from_file(config_model, tskey):
+    file_path = os.path.join(config_model.model.timeseries_data_path, tskey)
+    df = pd.read_csv(file_path, index_col=0)
+    return df
+
+
+def check_timeseries_dataframes(timeseries_dataframes):
+    """
+    Timeseries dataframes should be dict of pandas DataFrames.
+    """
+    if not isinstance(timeseries_dataframes, dict) or not all(
+        [
+            isinstance(timeseries_dataframes[i], pd.DataFrame)
+            for i in timeseries_dataframes
+        ]
+    ):
+        raise exceptions.ModelError(
+            "Error in loading timeseries data from dataframes. "
+            "`timeseries_dataframes` must be dict of pandas DataFrames."
+        )
+
+
+def load_timeseries_from_dataframe(timeseries_dataframes, tskey):
+
+    # If `df=` is called, timeseries_dataframes must be entered
+    if timeseries_dataframes is None:
+        raise exceptions.ModelError(
+            "Error in loading timeseries. Model config specifies df={} but "
+            "no timeseries passed as arguments in calliope.Model(...). "
+            "Note that, if running from a command line, it is not possible "
+            "to read dataframes via `df=...` and you should specify "
+            "`file=...` with a CSV file.".format(tskey)
+        )
+
+    try:
+        df = timeseries_dataframes[tskey]
+    except KeyError:
+        raise exceptions.ModelError(
+            "Error in loading data from dataframe. "
+            "Model attempted to load dataframe with key `{}`, "
+            "but time series passed as arguments are {}".format(
+                tskey, set(timeseries_dataframes.keys())
+            )
+        )
+    if not isinstance(df, pd.DataFrame):
+        raise exceptions.ModelError(
+            "Error in loading data. Object passed in time series "
+            "dictionary under key {} is a {}, not a DataFrame.".format(tskey, type(df))
+        )
+    return df
+
+
+def process_timeseries_data(config_model, model_run, timeseries_dataframes):
 
     if config_model.model.timeseries_data is None:
         timeseries_data = AttrDict()
@@ -443,51 +516,80 @@ def process_timeseries_data(config_model, model_run):
     def _parser(x, dtformat):
         return pd.to_datetime(x, format=dtformat, exact=False)
 
-    if "timeseries_data_path" in config_model.model:
-        dtformat = config_model.model["timeseries_dateformat"]
+    dtformat = config_model.model["timeseries_dateformat"]
 
-        # Generate the set of all files we want to read from file
-        location_config = model_run.locations.as_dict_flat()
-        model_config = config_model.model.as_dict_flat()
+    # Generate set of all files and dataframes we want to load
+    location_config = model_run.locations.as_dict_flat()
+    model_config = config_model.model.as_dict_flat()
 
-        get_filenames = lambda config: set(
+    # Find names of csv files (file=) or dataframes (df=) called in config
+    def get_names(datatype, config):
+        return set(
             [
                 v.split("=")[1].rsplit(":", 1)[0]
                 for v in config.values()
-                if "file=" in str(v)
+                if str(datatype) + "=" in str(v)
             ]
         )
-        constraint_filenames = get_filenames(location_config)
-        cluster_filenames = get_filenames(model_config)
 
-        _assert_timeseries_available(constraint_filenames | cluster_filenames)
+    constraint_filenames = get_names("file", location_config)
+    cluster_filenames = get_names("file", model_config)
+    constraint_dfnames = get_names("df", location_config)
+    cluster_dfnames = get_names("df", model_config)
 
-        datetime_min = []
-        datetime_max = []
+    # Check if timeseries_dataframes is in the correct format (dict of
+    # pandas DataFrames)
+    if timeseries_dataframes is not None:
+        check_timeseries_dataframes(timeseries_dataframes)
 
-        for file in constraint_filenames | cluster_filenames:
-            file_path = os.path.join(config_model.model.timeseries_data_path, file)
-            # load the data, without parsing the dates, to catch errors in the data
-            df = pd.read_csv(file_path, index_col=0)
-            try:
-                df.apply(pd.to_numeric)
-            except ValueError as e:
-                raise exceptions.ModelError(
-                    "Error in loading data from {}. Ensure all entries are "
-                    "numeric. Full error: {}".format(file, e)
-                )
-            # Now parse the dates, checking for errors specific to this
-            try:
-                df.index = _parser(df.index, dtformat)
-            except ValueError as e:
-                raise exceptions.ModelError(
-                    "Error in parsing dates in timeseries data from {}, "
-                    "using datetime format `{}`: {}".format(file, dtformat, e)
-                )
-            timeseries_data[file] = df
+    datetime_min = []
+    datetime_max = []
 
-            datetime_min.append(df.index[0].date())
-            datetime_max.append(df.index[-1].date())
+    # Check there is at least one timeseries present
+    if (
+        len(
+            constraint_filenames
+            | cluster_filenames
+            | constraint_dfnames
+            | cluster_dfnames
+        )
+        == 0
+    ):
+        raise exceptions.ModelError(
+            "There is no timeseries in the model. At least one timeseries is "
+            "necessary to run the model."
+        )
+
+    # Load each timeseries into timeseries data. tskey is either a filename
+    # (called by file=...) or a key in timeseries_dataframes (called by df=...)
+    for tskey in (
+        constraint_filenames | cluster_filenames | constraint_dfnames | cluster_dfnames
+    ):  # Filenames or dict keys
+        # If tskey is a CSV path, load the CSV, else load the dataframe
+        if tskey in constraint_filenames | cluster_filenames:
+            df = load_timeseries_from_file(config_model, tskey)
+        elif tskey in constraint_dfnames | cluster_dfnames:
+            df = load_timeseries_from_dataframe(timeseries_dataframes, tskey)
+
+        try:
+            df.apply(pd.to_numeric)
+        except ValueError as e:
+            raise exceptions.ModelError(
+                "Error in loading data from {}. Ensure all entries are "
+                "numeric. Full error: {}".format(tskey, e)
+            )
+        # Parse the dates, checking for errors specific to this
+        try:
+            df.index = _parser(df.index, dtformat)
+        except ValueError as e:
+            raise exceptions.ModelError(
+                "Error in parsing dates in timeseries data from {}, "
+                "using datetime format `{}`: {}".format(tskey, dtformat, e)
+            )
+        timeseries_data[tskey] = df
+
+        datetime_min.append(df.index[0].date())
+        datetime_max.append(df.index[-1].date())
 
     # Apply time subsetting, if supplied in model_run
     subset_time_config = config_model.model.subset_time
@@ -536,30 +638,24 @@ def process_timeseries_data(config_model, model_run):
 
     # Ensure all timeseries have the same index
     indices = [
-        (file, df.index)
-        for file, df in timeseries_data.items()
-        if file not in cluster_filenames
+        (tskey, df.index)
+        for tskey, df in timeseries_data.items()
+        if tskey not in cluster_filenames | cluster_dfnames
     ]
-    first_file, first_index = indices[0]
-    for file, idx in indices[1:]:
+    first_tskey, first_index = indices[0]
+    for tskey, idx in indices[1:]:
         if not first_index.equals(idx):
             raise exceptions.ModelError(
                 "Time series indices do not match "
-                "between {} and {}".format(first_file, file)
+                "between {} and {}".format(first_tskey, tskey)
             )
 
     return timeseries_data, first_index
 
 
-def _assert_timeseries_available(filenames):
-    if len(filenames) == 0:
-        raise exceptions.ModelError(
-            "There is no timeseries in the model. At least one timeseries is "
-            "necessary to run the model."
-        )
-
-
-def generate_model_run(config, debug_comments, applied_overrides, scenario):
+def generate_model_run(
+    config, timeseries_dataframes, debug_comments, applied_overrides, scenario
+):
     """
     Returns a processed model_run configuration AttrDict and a debug
     YAML object with comments attached, ready to write to disk.
@@ -567,8 +663,9 @@ def generate_model_run(config, debug_comments, applied_overrides, scenario):
     Parameters
     ----------
     config : AttrDict
+    timeseries_dataframes : dict
     debug_comments : AttrDict
-
+    scenario : str
     """
     model_run = AttrDict()
     model_run["scenario"] = scenario
@@ -600,7 +697,7 @@ def generate_model_run(config, debug_comments, applied_overrides, scenario):
     # 5) Fully populate timeseries data
     # Raises ModelErrors if there are problems with timeseries data at this stage
     model_run["timeseries_data"], model_run["timesteps"] = process_timeseries_data(
-        config, model_run
+        config, model_run, timeseries_dataframes
     )
 
     # 6) Grab additional relevant bits from run and model config