+++ changelog.rst
@@ -6,11 +6,17 @@ Release History
 0.6.6 (dev)
 -----------
 
+|new| Add ability to pass timeseries as dataframes in `calliope.Model` instead of only as CSV files.
+
+|new| Pyomo backend interfaces added to get names of all model objects (`get_all_model_attrs`) and to attach custom constraints to the backend model (`add_constraint`).
+
 |changed| Parameters are assigned a domain in Pyomo based on their dtype in `model_data`
 
 |changed| |backwards-incompatible| Plotting is no longer part of Calliope itself, but in a separate module, calliope_plot. FIXME: documentation needs to be updated.
 
 |changed| Internal code reorganisation.
+
+|changed| Updated to Pyomo 5.7.
 
 0.6.5 (2020-01-14)
 ------------------